% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mussel_barchart.R
\name{mussel_barchart}
\alias{mussel_barchart}
\title{Prints barcharts}
\usage{
mussel_barchart(data, x, y, ylabtext)
}
\arguments{
\item{y}{name of variable in river data to plot}
}
\value{
a plot
}
\description{
Prints barcharts
}
