% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mussel_boxplot.R
\name{mussel_boxplot}
\alias{mussel_boxplot}
\title{Prints boxplots by river}
\usage{
mussel_boxplot(y)
}
\arguments{
\item{y}{name of variable in river data to plot}
}
\value{
a plot
}
\description{
Prints boxplots by river
}
