% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mussel_boxplot.R
\name{mussel_boxplot_site}
\alias{mussel_boxplot_site}
\title{Prints boxplots by site}
\usage{
mussel_boxplot_site(y, species)
}
\arguments{
\item{y}{name of variable in river data to plot}
}
\value{
a plot
}
\description{
Prints boxplots by site
}
