% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mussel_plot.R
\name{mussel_plot}
\alias{mussel_plot}
\title{Prints plots of the mussels over time}
\usage{
mussel_plot(species, y)
}
\arguments{
\item{species}{name of species to plot}

\item{y}{name of variable in mussels to plot}
}
\value{
a plot
}
\description{
Prints plots of the mussels over time
}
