% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/river_plot.R
\name{river_plot}
\alias{river_plot}
\title{Prints plots of the river}
\usage{
river_plot(y)
}
\arguments{
\item{y}{name of variable in river data to plot}
}
\value{
a plot
}
\description{
Prints plots of the river
}
